Fix computation of dim colors on light background
Before, the dim colors were computed by scaling each individual color by
a factor of 0.66, thereby making them darker. That's correct for dark
backgrounds, but does not work on light backgrounds.

The implementation now takes the background into account and blends the
color into the background with an opacity of 0.5.